+++ packages/orq-rc/RELEASES.md
@@ -7199,3 +7199,13 @@ Based on:
 - [python v4.8.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.43] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.43 - packages/orq-rc
+
+## 2026-04-22 06:29:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.44] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.44] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.44 - packages/orq-rc